# fix(perf): Increase 8x speed of StringToLowerCase