This PR offers a significant performance optimization for string lowercase conversion functions.

## Problem
The current method using `CharToLower()` and `strlen()` is inefficient, especially for longer strings.

## Solution
Implementation of a new method using direct ASCII calculation that:
- Avoids calling `strlen()` which traverses the entire string
- Replaces `CharToLower()` with a simple mathematical operation
- Naturally stops at the null terminator character

## Benchmark Results
- Short strings: 526% faster
- Medium strings: 337% faster
- Long strings: 832% faster

## Testing
[A benchmark plugin was created ](https://pastecode.io/s/u5rs6mmt)to validate these improvements.

Complete results:
```cpp
===== BENCHMARK WITH SHORT STRING =====
Method 1 (CharToLower): 0.004428 seconds
Method 2 (ASCII math): 0.000707 seconds
Method 2 is 526.30% faster than Method 1 for short strings
===== BENCHMARK WITH MEDIUM STRING =====
Method 1 (CharToLower): 0.009437 seconds
Method 2 (ASCII math): 0.002156 seconds
Method 2 is 337.50% faster than Method 1 for medium strings
===== BENCHMARK WITH LONG STRING =====
Method 1 (CharToLower): 0.038702 seconds
Method 2 (ASCII math): 0.004151 seconds
Method 2 is 832.15% faster than Method 1 for long strings
```